requirements.txt
Streamline
Open
Elevenlabs
Python-dotenv
Pandas
Bumpy
